Realm backend works again now
The symbol table implementation I used originally in processMain started returning a null pointer.  As such, I changed to an implementation that iterates through the arguments of the function.  I feel this is ultimately more resilient anyway, because it doesn not rely on the arguments to main being named argc and argv respectively.